refactor(component): split chart-options-schema into per-chart files
Closes #258

Split the 1090-line monolithic chart-options-schema.ts into 20 focused files:
- chart-options/shared.ts — ChartOptionDef type, shared constants, reusable option arrays
- chart-options/{bar,line,pie,...}.ts — 17 per-chart option definitions
- chart-options/index.ts — registry, getChartOptions(), getDefaultChartSettings()
- chart-options-schema.ts — backward-compatible re-export shim

All 2,752 tests pass (1194 component + 1558 app). Zero API changes.
